requirements.txt
django-staticfiles
